+++ R/legend_fed.R
@@ -21,6 +21,8 @@
 #' when \code{ascii = TRUE}.
 #' 
 #' @param export (\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?
+#' 
+#' @param temp (\code{logical}). If \code{TRUE} keep temporary compressed file
 #'
 #' @details If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 #'  files in the current directory.
@@ -62,7 +64,11 @@
 #' df <- legend_fed(2002)
 #' }
 
-legend_fed <- function(year, uf = "all", br_archive = FALSE, ascii = FALSE, encoding = "latin1", export = FALSE){
+legend_fed <- function(year, uf = "all", br_archive = FALSE, 
+                       ascii = FALSE, 
+                       encoding = "latin1",
+                       export = FALSE,
+                       temp = TRUE){
 
 
   # Test the input
@@ -73,19 +79,25 @@ legend_fed <- function(year, uf = "all", br_archive = FALSE, ascii = FALSE, enco
   
   
   if(year < 2018) {
-    endereco <- "http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_legendas/consulta_legendas_%s.zip"
+    
+    endereco <- "http://cdn.tse.jus.br/estatistica/sead/odsele/consulta_legendas/consulta_legendas_%s.zip"
+    
+    filenames  <- paste0("/consulta_legendas_", year, ".zip")
+    
   } else{
-    endereco <- "http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_coligacao/consulta_coligacao_%s.zip"
+    endereco <- "http://cdn.tse.jus.br/estatistica/sead/odsele/consulta_coligacao/consulta_coligacao_%s.zip"
+    filenames  <- paste0("/consulta_coligacao_", year, ".zip")
   }
 
-  # Download the data
-  dados <- tempfile()
-  sprintf(endereco, year) %>%
-    download.file(dados)
-  unzip(dados, exdir = paste0("./", year))
-  unlink(dados)
+  dados <- paste0(file.path(tempdir()), filenames)
 
-  message("Processing the data...")
+  # Downloads the data
+  download_unzip(endereco, dados, filenames, year)
+  
+  # remover temp file
+  if(temp == FALSE){
+    unlink(dados)
+  }
 
   # Cleans the data
   setwd(as.character(year))
