+++ R/legend_local.R
@@ -16,6 +16,8 @@
 #' when \code{ascii = TRUE}.
 #' 
 #' @param export (\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?
+#' 
+#' @param temp (\code{logical}). If \code{TRUE} keep temporary compressed file
 #'
 #' @details If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 #'  files in the current directory.
@@ -56,7 +58,11 @@
 #' df <- legend_local(2000)
 #' }
 
-legend_local <- function(year, uf = "all", ascii = FALSE, encoding = "latin1", export = FALSE){
+legend_local <- function(year, uf = "all", 
+                         ascii = FALSE, 
+                         encoding = "latin1", 
+                         export = FALSE,
+                         temp = TRUE){
 
 
   # Test the input
@@ -65,14 +71,18 @@ legend_local <- function(year, uf = "all", ascii = FALSE, encoding = "latin1", e
   uf <- test_uf(uf)
   
   if(year > 2004){
-    # Download the data
-    dados <- tempfile()
-    sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_legendas/consulta_legendas_%s.zip", year) %>%
-      download.file(dados)
-    unzip(dados, exdir = paste0("./", year))
-    unlink(dados)
     
-    message("Processing the data...")
+    filenames  <- paste0("/consulta_legendas_", year, ".zip")
+    dados <- paste0(file.path(tempdir()), filenames)
+    url <- "https://cdn.tse.jus.br/estatistica/sead/odsele/consulta_legendas%s"
+    
+    # Downloads the data
+    download_unzip(url, dados, filenames, year)
+    
+    # remover temp file
+    if(temp == FALSE){
+      unlink(dados)
+    }
     
     # Cleans the data
     setwd(as.character(year))
