+++ R/seats_local.R
@@ -17,6 +17,8 @@
 #' when \code{ascii = TRUE}.
 #' 
 #' @param export (\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?
+#' 
+#' @param temp (\code{logical}). If \code{TRUE} keep temporary compressed file
 #'
 #' @details If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 #'  files in the current directory.
@@ -49,7 +51,10 @@
 #' df <- seats_local(2000)
 #' }
 
-seats_local <- function(year, uf = "all", ascii = FALSE, encoding = "latin1", export = FALSE){
+seats_local <- function(year, uf = "all", ascii = FALSE, 
+                        encoding = "latin1", 
+                        export = FALSE,
+                        temp = TRUE){
   
   
   # Input tests
@@ -57,14 +62,17 @@ seats_local <- function(year, uf = "all", ascii = FALSE, encoding = "latin1", ex
   test_local_year(year)
   uf <- test_uf(uf)
   
-  # Download the data
-  dados <- tempfile()
-  sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_vagas/consulta_vagas_%s.zip", year) %>%
-    download.file(dados)
-  unzip(dados, exdir = paste0("./", year))
-  unlink(dados)
+  filenames  <- paste0("/consulta_vagas_", year, ".zip")
+  dados <- paste0(file.path(tempdir()), filenames)
+  url <- "https://cdn.tse.jus.br/estatistica/sead/odsele/consulta_vagas%s"
   
-  message("Processing the data...")
+  # Downloads the data
+  download_unzip(url, dados, filenames, year)
+  
+  # remover temp file
+  if(temp == FALSE){
+    unlink(dados)
+  }
   
   # Cleans the data
   setwd(as.character(year))
